Introduce the class Drawer (#5)

+++ src/SixtyNine/Cloud/Placer/LissajouPlacer.php
@@ -33,6 +33,6 @@ class LissajouPlacer extends AbstractPlacer
     /** {@inheritdoc} */
     function getFirstPlaceToTry()
     {
-        return new Point($this->imgWidth / 2, $this->imgHeight / 2);
+        return new Point((int)($this->imgWidth / 2), (int)($this->imgHeight / 2));
     }
 }
